Reservation ID counter bug fixed. Gitignore amended to avoid commits of unnecessary files.
