+++ .gitignore
@@ -22,3 +22,12 @@
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
 replay_pid*
+
+*Other files
+.bin
+.md
+.gitattributes
+.iml
+.idea
+.name
+.xml
