+++ src/UI/AdminMenu/AdminMenuShowReservationsButton.java
@@ -1,24 +1,23 @@
-package UI.Admin;
-
-import DataTypesAndOperations.DataTypes.Customer;
-import Main.Main;
-import UI.MenuButton;
+package UI.AdminMenu;
 
 import java.io.IOException;
 
+import DataTypesAndOperations.DataTypes.Customer;
+import Main.*;
+import UI.*;
 
-public class AdminMenuDeleteCustomerButton implements MenuButton {
-    private String buttonText = "Delete an existing customer";
+
+public class AdminMenuShowReservationsButton implements MenuButton {
+    private String buttonText = "Show all registered reservations";
 
     @Override
     public void onPush() throws IOException {
-        Main.CUSTOMER_DELETER.start();
+        Main.RESERV_LIST_VIEWER.printList();
         Main.ADMIN.getToMenu();
     }
 
     @Override
     public void onPush(Customer customer) throws IOException {
-
     }
 
     @Override
