--- src/UI/Customer/CustMenuCancelReservButton.java
@@ -1,32 +1,0 @@
-package UI.Customer;
-
-import java.io.IOException;
-
-import DataTypesAndOperations.DataTypes.Customer;
-import Main.*;
-import UI.*;
-
-
-public class CustMenuCancelReservButton implements MenuButton {
-    private String buttonText = "Cancel my reservation";
-
-    @Override
-    public void onPush() {}
-
-    @Override
-    public void onPush(Customer customer) throws IOException {
-        Main.RESERV_CANCELLER.cancelReservation(customer);
-        customer.getToMenu();
-    }
-
-    @Override
-    public void show() {
-        Main.BUTTON_SHOWER.showButton(this, Main.CUST_MENU_BUTTONS);
-    }
-
-    @Override
-    public String getText() {
-        return buttonText;
-    }
-
-}
