+++ src/UI/CustomerMenu/CustMenuShowReservButton.java
@@ -1,4 +1,4 @@
-package UI.Customer;
+package UI.CustomerMenu;
 
 import java.io.IOException;
 
@@ -7,15 +7,16 @@ import Main.*;
 import UI.*;
 
 
-public class CustMenuCreateReservButton implements MenuButton {
-    private String buttonText = "Make a new reservation";
+public class CustMenuShowReservButton implements MenuButton {
+    private String buttonText = "Show my reservations";
 
     @Override
-    public void onPush() {}
+    public void onPush() {
+    }
 
     @Override
     public void onPush(Customer customer) throws IOException {
-        Main.RESERV_CREATOR.createReservation(customer);
+        Main.RESERV_LIST_VIEWER.printCustomerReservationsList(customer);
         customer.getToMenu();
     }
 
@@ -28,6 +29,4 @@ public class CustMenuCreateReservButton implements MenuButton {
     public String getText() {
         return buttonText;
     }
-
-
 }
